requirements.txt
pyTelegramBotAPI==4.15.0
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
nltk==3.8.1
python-dotenv==1.0.0
schedule==1.2.0
redis==5.0.1
gunicorn==21.2.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2